--- packages/core-v2-sdk/src/factories/Factory.ts
@@ -1,7 +1,0 @@
-// import { ElementClient } from "src/client";
-
-// export interface Factory<T> {
-//   client: ElementClient;
-//   get: (...args: any[]) => T | Promise<T>;
-//   getAll: (...args: any[]) => T[] | Promise<NonNullable<T>[]>;
-// }
