--- packages/core-v2-sdk/src/factories/YieldSourceFactory.ts
@@ -1,19 +1,0 @@
-// import { ElementClient } from "src/client";
-// import { YieldSource } from "src/models/YieldSource";
-// import { Factory } from "./Factory";
-
-// export class YieldSourceFactory implements Factory<YieldSource | null> {
-//   client: ElementClient;
-
-//   constructor(client: ElementClient) {
-//     this.client = client;
-//   }
-
-//   get(address: string): YieldSource | null {
-//     return new YieldSource({ address });
-//   }
-
-//   getAll(): YieldSource[] {
-//     return [];
-//   }
-// }
