+++ packages/core-v2-sdk/src/models/MultiPool.ts
@@ -1,16 +1,32 @@
+import { providers } from "ethers";
+import { ElementClient } from "src/client";
+import { MultiPoolContractDataSource } from "src/datasources/MultiPool/MultiPoolContractDataSource";
+import { MultiPoolDataSource } from "src/datasources/MultiPool/MultiPoolDataSource";
+import { setDataSourceByAddress } from "src/utils/setDataSourceByAddress";
 import { MultiTerm } from "./MultiTerm";
-
-export interface MultiPoolFields {
-  address: string;
-  multiTerm: MultiTerm;
-}
 
 export class MultiPool {
   address: string;
-  multiTerm: MultiTerm;
+  dataSource: MultiPoolDataSource;
+  provider: providers.BaseProvider;
 
-  constructor({ address, multiTerm }: MultiPoolFields) {
+  constructor(
+    address: string,
+    client: ElementClient,
+    dataSource?: MultiPoolDataSource,
+  ) {
     this.address = address;
-    this.multiTerm = multiTerm;
+    this.provider = client.provider;
+    this.dataSource =
+      dataSource ??
+      setDataSourceByAddress(
+        new MultiPoolContractDataSource({ address, provider: client.provider }),
+        client,
+      );
+  }
+
+  async getTerm(): Promise<MultiTerm> {
+    const address = await this.dataSource.getMultiTerm();
+    return new MultiTerm(address, this.provider);
   }
 }
