+++ packages/core-v2-sdk/src/utils/setDataSourceByAddress.ts
@@ -1,0 +1,13 @@
+import { ElementClient } from "src/client";
+import { getDataSourceByAddress } from "./getDataSourceByAddress";
+
+export function setDataSourceByAddress<T extends { address: string }>(
+  dataSource: T,
+  client: ElementClient,
+): T {
+  if (!getDataSourceByAddress(dataSource.address, client)) {
+    client.dataSources.push(dataSource);
+  }
+
+  return dataSource;
+}
